requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=conda_forge
_openmp_mutex=4.5=1_llvm
absl-py=0.13.0=pyhd8ed1ab_0
aiohttp=3.7.4.post0=py39h3811e60_0
albumentations=1.0.3=pypi_0
alsa-lib=1.2.3=h516909a_0
argparse=1.4.0=pypi_0
async-timeout=3.0.1=py_1000
attrs=21.2.0=pyhd8ed1ab_0
backcall=0.2.0=pypi_0
batchgenerators=0.23=pypi_0
blas=1.0=mkl
blinker=1.4=py_1
blosc=1.21.0=h8c45485_0
bottleneck=1.3.2=py39hdd57654_1
brotli=1.0.9=he6710b0_2
brotlipy=0.7.0=py39h3811e60_1001
brunsli=0.1=h2531618_0
bzip2=1.0.8=h7b6447c_0
c-ares=1.17.2=h7f98852_0
ca-certificates=2021.7.5=h06a4308_1
cached-property=1.5.2=hd8ed1ab_1
cached_property=1.5.2=pyha770c72_1
cachetools=4.2.2=pyhd8ed1ab_0
cairo=1.16.0=h6cf1ce9_1008
certifi=2021.5.30=py39h06a4308_0
cffi=1.14.6=py39he32792d_0
cfitsio=3.470=hf0d0db6_6
chardet=4.0.0=py39hf3d152e_1
charls=2.2.0=h2531618_0
charset-normalizer=2.0.0=pyhd8ed1ab_0
click=8.0.1=py39hf3d152e_0
cloudpickle=1.6.0=py_0
colorama=0.4.4=pyh9f0ad1d_0
cryptography=3.4.7=py39hbca0aa6_0
cudatoolkit=10.2.89=hfd86e86_1
cycler=0.10.0=py39h06a4308_0
cytoolz=0.11.0=py39h27cfd23_0
dask-core=2021.8.0=pyhd3eb1b0_0
dataclasses=0.8=pyhc8e2a94_3
dbus=1.13.6=h48d8840_2
decorator=5.1.0=pypi_0
dicom2nifti=2.3.0=pypi_0
expat=2.4.1=h9c3ff4c_0
ffmpeg=4.3.2=hca11adc_0
fontconfig=2.13.1=hba837de_1005
fonttools=4.25.0=pyhd3eb1b0_0
freetype=2.10.4=h5ab3b9f_0
fsspec=2021.7.0=pyhd8ed1ab_0
future=0.18.2=py39hf3d152e_3
gettext=0.19.8.1=h0b5b191_1005
giflib=5.1.4=h14c3975_1
glib=2.68.4=h9c3ff4c_0
glib-tools=2.68.4=h9c3ff4c_0
gmp=6.2.1=h58526e2_0
gnutls=3.6.13=h85f3911_1
google-auth=1.35.0=pyh6c4a22f_0
google-auth-oauthlib=0.4.1=py_2
graphite2=1.3.13=h58526e2_1001
grpcio=1.38.1=py39hff7568b_0
gst-plugins-base=1.18.4=hf529b03_2
gstreamer=1.18.4=h76c114f_2
h5py=3.2.1=py39h6c542dc_0
harfbuzz=2.9.0=h83ec7ef_0
hdf5=1.10.6=hb1b8bf9_0
hiddenlayer=0.2=pypi_0
icu=68.1=h58526e2_0
idna=3.1=pyhd3deb0d_0
imagecodecs-lite=2019.12.3=py39hce5d2b2_3
imageio=2.9.0=pyhd3eb1b0_0
importlib-metadata=4.6.4=py39hf3d152e_0
ipython=7.27.0=pypi_0
jasper=1.900.1=h07fcdf6_1006
jedi=0.18.0=pypi_0
joblib=1.0.1=pyhd3eb1b0_0
jpeg=9d=h36c2ea0_0
jxrlib=1.1=h7b6447c_2
kiwisolver=1.3.1=py39h2531618_0
krb5=1.19.2=hcc1bbae_0
lame=3.100=h7f98852_1001
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.36.1=hea4e1c9_2
lerc=2.2.1=h2531618_0
libaec=1.0.4=he6710b0_1
libblas=3.9.0=8_mkl
libcblas=3.9.0=8_mkl
libclang=11.1.0=default_ha53f305_1
libcurl=7.78.0=h2574ce0_0
libdeflate=1.7=h27cfd23_5
libedit=3.1.20191231=he28a2e2_2
libev=4.33=h516909a_1
libevent=2.1.10=hcdb4288_3
libffi=3.3=h58526e2_2
libgcc-ng=11.1.0=hc902ee8_8
libgfortran-ng=7.5.0=ha8ba4b0_17
libgfortran4=7.5.0=ha8ba4b0_17
libglib=2.68.4=h3e27bee_0
libiconv=1.16=h516909a_0
liblapack=3.9.0=8_mkl
liblapacke=3.9.0=8_mkl
libllvm11=11.1.0=hf817b99_2
libnghttp2=1.43.0=h812cca2_0
libogg=1.3.4=h7f98852_1
libopencv=4.5.1=py39hf3c14e0_0
libopus=1.3.1=h7f98852_1
libpng=1.6.37=hbc83047_0
libpq=13.3=hd57d9b9_0
libprotobuf=3.15.8=h780b84a_0
libssh2=1.9.0=ha56f1ee_6
libstdcxx-ng=11.1.0=h56837e0_8
libtiff=4.2.0=h85742a9_0
libuuid=2.32.1=h7f98852_1000
libuv=1.40.0=h7b6447c_0
libvorbis=1.3.7=h9c3ff4c_0
libwebp-base=1.2.1=h7f98852_0
libxcb=1.13=h7f98852_1003
libxkbcommon=1.0.3=he3ba5ed_0
libxml2=2.9.12=h72842e0_0
libzopfli=1.0.3=he6710b0_0
linecache2=1.0.0=pypi_0
llvm-openmp=12.0.1=h4bd325d_1
locket=0.2.1=py39h06a4308_1
lz4-c=1.9.3=h295c915_1
markdown=3.3.4=pyhd8ed1ab_0
matplotlib-base=3.4.2=py39hab158f2_0
matplotlib-inline=0.1.3=pypi_0
medpy=0.4.0=pypi_0
mkl=2020.4=h726a3e6_304
mkl-service=2.3.0=py39he8ac12f_0
monai=0.6.0=pypi_0
multidict=5.1.0=py39h3811e60_1
munkres=1.1.4=py_0
mysql-common=8.0.25=ha770c72_0
mysql-libs=8.0.25=h935591d_0
ncurses=6.2=h58526e2_4
nettle=3.6=he412f7d_0
networkx=2.6.2=pyhd3eb1b0_0
nibabel=3.2.1=pyhd8ed1ab_0
nii2png=0.2.7=pypi_0
ninja=1.10.2=h4bd325d_0
nnunet=1.7.0=dev_0
nspr=4.30=h9c3ff4c_0
nss=3.69=hb5efdd6_0
numexpr=2.7.3=py39hb2eb853_0
numpy=1.21.2=py39hdbf815f_0
oauthlib=3.1.1=pyhd8ed1ab_0
olefile=0.46=py_0
opencv=4.5.1=py39hf3d152e_0
opencv-python-headless=4.5.3.56=pypi_0
openh264=2.1.1=h780b84a_0
openjpeg=2.3.0=h05c96fa_1
openssl=1.1.1l=h7f8727e_0
packaging=21.0=pyhd8ed1ab_0
pandas=1.3.2=py39h8c16a72_0
parso=0.8.2=pypi_0
partd=1.2.0=pyhd3eb1b0_0
pcre=8.45=h9c3ff4c_0
pexpect=4.8.0=pypi_0
pickleshare=0.7.5=pypi_0
pillow=8.3.1=py39h2c7a002_0
pip=21.2.4=pypi_0
pixman=0.40.0=h36c2ea0_0
prompt-toolkit=3.0.20=pypi_0
protobuf=3.15.8=py39he80948d_0
pthread-stubs=0.4=h36c2ea0_1001
ptyprocess=0.7.0=pypi_0
py-opencv=4.5.1=py39hef51801_0
pyasn1=0.4.8=py_0
pyasn1-modules=0.2.7=py_0
pycparser=2.20=pyh9f0ad1d_2
pydeprecate=0.3.1=pyhd8ed1ab_0
pydicom=2.2.0=pyh6c4a22f_0
pygments=2.10.0=pypi_0
pyjwt=2.1.0=pyhd8ed1ab_0
pyopenssl=20.0.1=pyhd8ed1ab_0
pyparsing=2.4.7=pyh9f0ad1d_0
pypng=0.0.20=py_0
pysocks=1.7.1=py39hf3d152e_3
python=3.9.6=h49503c6_1_cpython
python-dateutil=2.8.2=pyhd3eb1b0_0
python_abi=3.9=2_cp39
pytorch=1.9.0=py3.9_cuda10.2_cudnn7.6.5_0
pytorch-lightning=1.4.6=pyhd8ed1ab_0
pytz=2021.1=pyhd3eb1b0_0
pyu2f=0.1.5=pyhd8ed1ab_0
pywavelets=1.1.1=py39h6323ea4_4
pyyaml=5.4.1=py39h3811e60_1
qt=5.12.9=hda022c4_4
readline=8.1=h46c0cb4_0
requests=2.26.0=pyhd8ed1ab_0
requests-oauthlib=1.3.0=pyh9f0ad1d_0
rsa=4.7.2=pyh44b312d_0
scikit-image=0.18.1=py39ha9443f7_0
scikit-learn=0.24.2=py39ha9443f7_0
scipy=1.6.2=py39h91f5cce_0
setuptools=57.4.0=pypi_0
simpleitk=2.1.0=py39h3fd9d12_0
six=1.16.0=pyh6c4a22f_0
sklearn=0.0=pypi_0
sleef=3.5.1=h7f98852_1
snappy=1.1.8=he6710b0_0
sqlite=3.36.0=h9cd32fc_0
tensorboard=2.6.0=pyhd8ed1ab_0
tensorboard-data-server=0.6.0=py39h3da14fd_0
tensorboard-plugin-wit=1.8.0=pyh44b312d_0
threadpoolctl=2.2.0=pyh0d69192_0
tifffile=2020.6.3=py_0
tk=8.6.10=h21135ba_1
toolz=0.11.1=pyhd3eb1b0_0
torchaudio=0.9.0=py39
torchmetrics=0.5.0=pyhd8ed1ab_0
torchvision=0.10.0a0+300a8a4=pypi_0
tornado=6.1=py39h27cfd23_0
tqdm=4.62.1=pyhd8ed1ab_0
traceback2=1.4.0=pypi_0
traitlets=5.1.0=pypi_0
typing-extensions=3.10.0.0=hd8ed1ab_0
typing_extensions=3.10.0.0=pyha770c72_0
tzdata=2021a=he74cb21_1
unittest2=1.1.0=pypi_0
urllib3=1.26.6=pyhd8ed1ab_0
wcwidth=0.2.5=pypi_0
werkzeug=2.0.1=pyhd8ed1ab_0
wheel=0.37.0=pyhd8ed1ab_1
x264=1!161.3030=h7f98852_1
xorg-kbproto=1.0.7=h7f98852_1002
xorg-libice=1.0.10=h7f98852_0
xorg-libsm=1.2.3=hd9c2040_1000
xorg-libx11=1.7.2=h7f98852_0
xorg-libxau=1.0.9=h7f98852_0
xorg-libxdmcp=1.1.3=h7f98852_0
xorg-libxext=1.3.4=h7f98852_1
xorg-libxrender=0.9.10=h7f98852_1003
xorg-renderproto=0.11.1=h7f98852_1002
xorg-xextproto=7.3.0=h7f98852_1002
xorg-xproto=7.0.31=h7f98852_1007
xz=5.2.5=h516909a_1
yaml=0.2.5=h516909a_0
yarl=1.6.3=py39h3811e60_2
zfp=0.5.5=h2531618_6
zipp=3.5.0=pyhd8ed1ab_0
zlib=1.2.11=h516909a_1010
zstd=1.4.9=haebb681_0
